Final changes after review + update on square mode detection
